refactor:  remove diseaseName from data layer
