+++ src/domain/repositories/PerformanceOverviewRepository.ts
@@ -1,5 +1,4 @@
 import { FutureData } from "../../data/api-futures";
-import { Maybe } from "../../utils/ts-utils";
 import { AlertsPerformanceOverviewMetrics } from "../entities/alert/AlertsPerformanceOverviewMetrics";
 import {
     DataSource,
@@ -9,7 +8,6 @@ import {
     TotalCardCounts,
     PerformanceOverviewMetrics,
     PerformanceMetrics717,
-    DiseaseNames,
 } from "../entities/disease-outbreak-event/PerformanceOverviewMetrics";
 import { OverviewCard } from "../entities/PerformanceOverview";
 import { Id } from "../entities/Ref";
@@ -28,7 +26,7 @@ export interface PerformanceOverviewRepository {
     ): FutureData<TotalCardCounts[]>;
     getNational717Performance(): FutureData<PerformanceMetrics717[]>;
     getEvent717Performance(diseaseOutbreakEventId: Id): FutureData<PerformanceMetrics717[]>;
-    getAlerts717Performance(diseaseName: Maybe<DiseaseNames>): FutureData<PerformanceMetrics717[]>;
+    getAlerts717Performance(): FutureData<PerformanceMetrics717[]>;
     getEventTrackerOverviewMetrics(
         type: string,
         dataSource?: DataSource
